refactor: move data/local/remote dir to each module
