+++ core/local/src/main/java/com/emotionstorage/local/di/LocalDataSourceModule.kt
@@ -1,9 +1,11 @@
 package com.emotionstorage.local.di
 
+import com.emotionstorage.data.dataSource.local.AttendanceLocalDataSource
 import com.emotionstorage.data.dataSource.local.FcmLocalDataSource
 import com.emotionstorage.data.dataSource.local.SessionLocalDataSource
 import com.emotionstorage.data.dataSource.local.TimeCapsuleLocalDataSource
 import com.emotionstorage.data.dataSource.local.UserLocalDataSource
+import com.emotionstorage.local.dataSourceImpl.AttendanceLocalDataSourceImpl
 import com.emotionstorage.local.dataSourceImpl.FcmLocalDataSourceImpl
 import com.emotionstorage.local.dataSourceImpl.SessionLocalDataSourceImpl
 import com.emotionstorage.local.dataSourceImpl.TimeCapsuleLocalDataSourceImpl
@@ -17,6 +19,10 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class LocalDataSourceModule {
+
+    @Binds
+    abstract fun bindLocalDataSource(impl: AttendanceLocalDataSourceImpl): AttendanceLocalDataSource
+
     @Binds
     @Singleton
     abstract fun bindFcmLocalDataSource(impl: FcmLocalDataSourceImpl): FcmLocalDataSource
