+++ core/remote/src/main/java/com/emotionstorage/remote/api/HomeApiService.kt
@@ -1,12 +1,12 @@
-package com.emotionstorage.home.remote.api
+package com.emotionstorage.remote.api
 
-import com.emotionstorage.home.remote.response.HomeResponseData
 import com.emotionstorage.remote.interceptor.AuthRequest
+import com.emotionstorage.remote.response.home.HomeResponse
 import com.emotionstorage.remote.response.ResponseDto
 import retrofit2.http.GET
 
 interface HomeApiService {
     @AuthRequest
     @GET("/api/v1/home")
-    suspend fun getHome(): ResponseDto<HomeResponseData>
+    suspend fun getHome(): ResponseDto<HomeResponse>
 }
