+++ core/remote/src/main/java/com/emotionstorage/remote/di/ApiServiceModule.kt
@@ -2,6 +2,7 @@ package com.emotionstorage.remote.di
 
 import com.emotionstorage.remote.api.DailyReportApiService
 import com.emotionstorage.remote.api.FcmApiService
+import com.emotionstorage.remote.api.HomeApiService
 import com.emotionstorage.remote.api.MyPageApiService
 import com.emotionstorage.remote.api.TimeCapsuleApiService
 import com.emotionstorage.remote.api.UserApiService
@@ -15,6 +16,11 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object ApiServiceModule {
+
+    @Singleton
+    @Provides
+    fun provideHomeApiService(retrofit: Retrofit): HomeApiService = retrofit.create(HomeApiService::class.java)
+
     @Singleton
     @Provides
     fun provideFcmApiService(retrofit: Retrofit): FcmApiService = retrofit.create(FcmApiService::class.java)
