+++ core/remote/src/main/java/com/emotionstorage/remote/response/home/HomeResponse.kt
@@ -1,9 +1,9 @@
-package com.emotionstorage.home.remote.response
+package com.emotionstorage.remote.response.home
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HomeResponseData(
+data class HomeResponse(
     // 보유 티켓 수
     val remainingTickets: Int,
     // 하루 최대 보유 가능한 티켓 수
