+++ data/src/main/java/com/emotionstorage/data/dataSource/remote/HomeRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.emotionstorage.data.dataSource.remote
+
+import com.emotionstorage.data.model.HomeEntity
+
+interface HomeRemoteDataSource {
+    suspend fun getHome(): HomeEntity
+}
