+++ feat/home/build.gradle.kts
@@ -2,8 +2,6 @@ plugins {
     id("com.emotionstorage.convention.android.library")
     id("com.emotionstorage.convention.android.library.compose")
     id("com.emotionstorage.convention.android.library.hilt")
-    id("com.emotionstorage.convention.kotlin.library.retrofit")
-    id("com.emotionstorage.convention.android.library.datastore")
 }
 
 android {
