--- feat/home/src/main/java/com/emotionstorage/home/data/dataSource/HomeRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.emotionstorage.home.data.dataSource
-
-import com.emotionstorage.home.data.model.HomeEntity
-
-interface HomeRemoteDataSource {
-    suspend fun getHome(): HomeEntity
-}
