--- feat/home/src/main/java/com/emotionstorage/home/remote/di/ApiServiceModule.kt
@@ -1,17 +1,0 @@
-package com.emotionstorage.home.remote.di
-
-import com.emotionstorage.home.remote.api.HomeApiService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ApiServiceModule {
-    @Singleton
-    @Provides
-    fun provideHomeApiService(retrofit: Retrofit): HomeApiService = retrofit.create(HomeApiService::class.java)
-}
